# pgtype/hstore: Avoid Postgres Mac OS X parsing bug